+++ clang-tools-extra/test/clang-tidy/checkers/readability/redundant-member-init.cpp
@@ -318,3 +318,24 @@ struct D9 {
   D9() : ss(S()) {}
   SS ss;
 };
+
+#define EMPTY
+#define BRACES {}
+
+struct WithMacro1 {
+  S s BRACES;
+  // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: initializer for member 's' is redundant
+};
+
+struct WithoutMacroCtor {
+  WithoutMacroCtor() : s(EMPTY) {};
+  // CHECK-MESSAGES: :[[@LINE-1]]:24: warning: initializer for member 's' is redundant
+  // CHECK-FIXES: WithoutMacroCtor()  {};
+  S s;
+};
+
+struct WithBlockCommentInInit {
+  S s {/* default-construct */};
+  // CHECK-MESSAGES: :[[@LINE-1]]:7: warning: initializer for member 's' is redundant
+  // CHECK-FIXES: S s;
+};